[core] move RenderItem and GlyphRange to their own headers
