+++ src/mbgl/text/glyph_range.hpp
@@ -1,0 +1,13 @@
+#ifndef MBGL_TEXT_GLYPH_RANGE
+#define MBGL_TEXT_GLYPH_RANGE
+
+#include <utility>
+#include <cstdint>
+
+namespace mbgl {
+
+typedef std::pair<uint16_t, uint16_t> GlyphRange;
+
+} // end namespace mbgl
+
+#endif
